Some changes to comply with MSVC compiler

+++ src/fontio/logic/type2/Type2SubroutineAccessor.hpp
@@ -35,7 +35,7 @@ namespace fontio { namespace logic { namespace type2
             }
 
             auto unbiased = this->UnbiasIndex(index);
-            if ((unbiased < 0) || (unbiased >= this->subroutines->size()))
+            if ((unbiased < 0) || (unbiased >= static_cast<int>(this->subroutines->size())))
             {
                 throw std::runtime_error("Index out of range");
             }
